+++ src/main/java/com/cognixia/jumplus/project3/anorthouse/service/LogOut.java
@@ -1,4 +1,4 @@
-package com.cognixia.jumplus.project3.anorthouse.web;
+package com.cognixia.jumplus.project3.anorthouse.service;
 
 import java.io.IOException;
 import jakarta.servlet.ServletException;
